Refactor GraphicsDeviceInformation generation into a common function.
It turns out that IGraphicsDeviceManager.CreateDevice actually destroys previous
devices, so this made for a really annoying problem where CreateDevice and
ApplyChanges couldn't be circularly-dependent. This ended up revealing _tons_ of
accuracy failures in PreparingDeviceSettings and the various GDM settings!

Now both ApplyChanges and CreateDevice effectively behave the same, except now
CreateDevice exclusively creates the device and ApplyChanges exclusively resets
the device. And yes, ApplyChanges in the constructor was actually a bad idea.