In FNADllMap, prefer AppContext.BaseDirectory
IL3000: Assembly.Location always returns an empty string for assemblies
embedded in a single-file app. Use AppContext.BaseDirectory instead.

Upon investigation, AppContext.BaseDirectory seems to be exactly the
path that's needed here in all cases, and it is reported to work on all
versions of .net core and .net framework.