Replace Marshal.SizeOf(typeof(T)) with the generic version
I considered replacing it with just regular sizeof(T), but sizeof(T)
returns the *managed* size, and this call really does want the
*unmanaged* size. I think for all of these structs those may be the
same, but just in case there is a bool or a char in them, I stuck with
Marshal.SizeOf.

In order to continue to support .NET Framework 4.0, we introduce a
helper to still call the reflection-based version there.